update permute_last2dim file
